functions/src/requirements.txt
# Core dependencies with latest versions
flask>=2.1.2
functions-framework>=3.0.0
google-cloud-storage>=2.1.0
firebase-admin>=6.1.0
stripe>=8.0.0
requests>=2.27.1
python-dotenv>=0.19.2
gunicorn>=20.1.0
google-cloud-aiplatform>=1.36.3
PyJWT>=2.8.0

# Flask-compatible dependencies
werkzeug>=2.1.2
markupsafe>=2.0.1
jinja2>=3.0.3
itsdangerous>=2.0.1
click>=8.0.3

# Keep using Pydantic v1 (this matches your code after applying the fix to auth.py)
pydantic>=1.10.12
typing-extensions>=4.2.0

# Agent dependencies
langgraph>=0.0.19
langchain>=0.1.0
langchain-core>=0.1.22
langchain-google-genai>=0.0.9
google-generativeai>=0.3.2
aiohttp>=3.9.3
markdown2>=2.4.12  # For Markdown processing

# PDF generation - using a pure Python library that works in Cloud Functions
xhtml2pdf>=0.2.11  # Pure Python HTML to PDF converter, no system dependencies